Renaming modules to match PluralSight.
Fixing one test
