+++ tasks_module2.md
@@ -1,89 +1,65 @@
----
-% Django Models - Blog Posts Model
-% Will Holderness
----
+# Create user app
 
+We’re going to create a separate user app within our project.  Run the following command to do that:  
+`python manage.py startapp users`
 
-# Create Model class
+# Include the new app into the project
 
-Create the blog post model by adding a new class into `mainapp/models.py`. Name the class `BlogPost` and make it extend `models.Model`. In the class body, add a field called `title` and set it to `models.CharField()`. Pass it a `max_length` of `200`.
+Inside of `settings.py` (located in `blogproj/`), find the `INSTALLED_APPS` list and after `'mainapp'`, add `'users'`.
 
-```python
-class BlogPost(models.Model):
-    title = models.CharField(max_length=200)
+# Add login template
+
+We went ahead and created a basic login template for you to copy.  First, create a folder inside the `users/` directory called `templates`. Then, from the `support/` directory, move the `login.html` file into the `users/templates/` directory.
+
+# Create custom user model
+
+Now we want to create our User model in `users/models.py`.  To create a User class that derives from AbstractUser do the following:
+Add an import of `AbstractUser` from `django.contrib.auth.models` in the import section.
+Create the class `User` and derive from `AbstractUser` like so -- 
+`class User(AbstractUser):`
+Leave the class definition empty with `pass`, the base class has all we need for now.
+
+# Add user model to the admin site
+
+In order to have access to this model in the admin, we need to register it in `admin.py`.  Let’s do the following:
+In the import section, add an import of `UserAdmin` from `django.contrib.auth.admin` and import `User` from `.models`.
+Call `admin.site.register()` with `User` and `UserAdmin` as parameters.
+
+# Set user settings in the main project
+
+We need to specify the login URLs and the User model to use for authentication in `settings.py`.  At the bottom of `settings.py` in the `blogproj/` directory, add three global variables. First add `LOGIN_URL` and set it to `'login'`. Next, add `LOGIN_REDIRECT_URL` and set it to `'index'`. Lastly, add `AUTH_USER_MODEL` with the value of  `'users.User'`
+
+# Add routes to the user app
+
+Next, we need to setup routes for the user project. Like in the main app, we’ll create a dedicated `urls.py` file inside of the `users` directory.   We’re going to use `LoginView.as_view()` and `LogoutView.as_view()` to provide built-in login and logout functionality.  In the newly created `urls.py` file, import `views` as `auth_views`from `from django.contrib.auth`, import `path` from `django.urls`, and import `views` from `.` (the current folder). Then copy and paste the following URL patterns at the bottom of the file:
+```
+urlpatterns = [
+    path('login/', auth_views.LoginView.as_view(template_name='login.html'), name='login'),
+    path('logout/', auth_views.LogoutView.as_view(next_page='index' ), name='logout'),
+]
 ```
 
-# Add author foreign key
+# Redirect user related routes to the user project
 
-To add the foreign key, we need to know which user model is in use. This is stored in the Django setting called `AUTH_USER_MODEL`. Lets first import those settings.
+In `blogproj/urls.py`, add `path('accounts/', include('users.urls')),` as the first entry in the `urlpatterns` array.
 
-Add `from django.conf import settings` to the top of the file. Then in the body of our `BlogPost` class, add a new field called `author` with type of `models.ForeignKey()`. Pass the following parameters to `ForeignKey()`: `settings.AUTH_USER_MODEL`, `related_name="posts"`, and `on_delete=models.CASCADE`.
-    
-```python
-author = models.ForeignKey(
-    settings.AUTH_USER_MODEL,
-    related_name="posts",
-    on_delete=models.CASCADE,
-)
-```
+# Add login/logout to site templates
 
-# Add body field
+In the `templates/base.html` template, uncomment the `Login` and `Logout` list items by removing the `{% comment %}` and `{% uncomment %}` tags inside the `navbar-nav` div.
 
-Add another field to `BlogPost` called `body` with type `models.TextField()`
+# Run initial migrations
 
-# Add a date field to each post
-
-Add a field called `postdate` of type `models.DateTimeField()` with parameters `auto_now_add` set to `True`, and `blank` set to `True`.
-
-`postdate = models.DateTimeField(auto_now_add=True, blank=True)`
-
-# Add string method
-Adding a `__str__()` method to the class will simplify rendering in admin view and our templates. In the `BlogPost` class, create a method called `__str__` that takes a parameter called `self`. Have the method return `self.title`. 
-
-```python
-def __str__(self):
-	return self.title
-```
-
-# Add url reverser.
-
-Having a `get_absolute_url` function lets Django determine the canonical URL for a given model. This will come in handy in our views. First we need to import the `reverse` helper function from `django.urls` at the top of the file. Next add a method in the class called `get_absolute_url` that takes the `self` parameter. The method should contain the following line: `return reverse('post', args=[str(self.id)])`
-
-```python
-def get_absolute_url(self):
-	return reverse('post', args=[str(self.id)])
-```
-
-# Register new model on admin site
-
-In `mainapp/admin.py`, import the new `BlogPost` model from `.models`. Then at the bottom of the file, call `admin.site.register()` with `BlogPost` as the only parameter.
-
-```python
-from .models import BlogPost
-
-admin.site.register(BlogPost)
-```
-
-# Update index view to use model
-
-Now we need to update `mainapp/views.py` to include our new `BlogPost` model. First let's import the `BlogPost` model from `.models`. Also import the `get_object_or_404` method from `django.shortcuts`.
-
-Next in the `index` method, set the `posts` variable to `BlogPost.objects.all()` instead of `ALL_POSTS`.
-
-# Include model in post view
-
-In the `post` method, we can take advantage of the django shortcut `get_object_or_404()` to do all the exception handling instead of having to rewrite it often. Replace the entire try/accept block with a line that sets the `post` variable to the BlogPost returned by `get_object_or_404(BlogPost, pk=id)` . 
-The `return` line can stay just as it is.
-
-
-# Migrate to new schema
-
-Now that we’ve added our `BlogPost` model, we need to run migrations again. From the command line, inside the root of the project, run the following commands:
-
+Now that we’ve added our User model we need to run migrations. From the command line, inside the root of the project, run the following commands:
 `python manage.py makemigrations`
 `python manage.my migrate`
 
-Once created, add the newly created migrations from `mainapp/migrations/` to the git repo and commit them.
+Once created, add the newly created migrations to the git repo and commit them.
 
-Note: Make sure your `venv` is activated with `source venv/bin/activate`
- 
+**Note:** Make sure your `venv` is activated with `source venv/bin/activate`
+**Note:** If you made migrations before adding the user model, you’ll want to remove them because it will cause errors. You can do this by deleting all files in the `mainapp/migrations` folder, `users/migrations` folder, and the `db.sqlite3` file.
+
+# Create an admin user
+
+Create a superuser by running the following command from the command line: `python manage.py createsuperuser`.
+
+At this point we can login and logout with our superuser.  Let’s test this out by running `python manage.py runserver`, then visit `http://localhost:8000`.  Click on the Login link, then login with your superuser credentials.  You can also login to the admin at `http://localhost:8000/admin`.  Congratulations!  You successfully implemented a custom User model with authentication.
